release version 1.6.8 - update kotlin version to 1.8.21 - update inner libs - update gradle to 8.1.1 and agp to 7.4.2

+++ app/src/main/java/com/mincor/kodiexample/data/model/local/MovieEntity.kt
@@ -11,7 +11,7 @@ import com.mincor.kodiexample.presentation.movies.MovieUI
 data class MovieEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Int,
-    val voteCount: Int,
+    val voteCount: Long,
     val voteAverage: Double,
     val isVideo: Boolean,
     val title: String,
@@ -24,8 +24,8 @@ data class MovieEntity(
     val releaseDate: Long,
     val adult: Boolean,
     val overview: String,
-    val revenue: Int,
-    val runtime: Int,
+    val revenue: Long,
+    val runtime: Long,
     val status: String,
     val tagline: String,
     val budget: Long,
